WEBAPP-215: Rename to CategoryTimeStamp
