+++ src/routes/categories/components/CategoryTimeStamp.js
@@ -4,7 +4,7 @@ import React from 'react'
 import CategoryModel from '../../../modules/endpoint/models/CategoryModel'
 import type { I18nTranslate } from '../../../flowTypes'
 import { translate } from 'react-i18next'
-import { Timestamp } from './CategoryTimestamp.styles'
+import { Timestamp } from './CategoryTimeStamp.styles'
 
 type Props = {
   language: string,
@@ -12,7 +12,7 @@ type Props = {
   t: I18nTranslate
 }
 
-class CategoryTimestamp extends React.PureComponent<Props> {
+class CategoryTimeStamp extends React.PureComponent<Props> {
   render () {
     const {language, category, t} = this.props
 
@@ -20,4 +20,4 @@ class CategoryTimestamp extends React.PureComponent<Props> {
   }
 }
 
-export default translate('categories')(CategoryTimestamp)
+export default translate('categories')(CategoryTimeStamp)
